[FIX] sale: correctly test all files
The rev. odoo/odoo@aad674f has recently been forward-ported but was in conflict
with the rev odoo/odoo@4404d5d ; both have created a inherited view to add their
test file in the qunit_suite view.

As both inherited views have the same id, one of the two files was not correctly
loaded in the test suite.

closes odoo/odoo#29586