[FIX] stock: wrong location technical field
The commit 4c2f906b9781d142b9d2175920b029e9e34bb684 changes the odoo object
of "rules" from stock.location.path to stock.rule. The field on the first
object to store the destination location was location_dest_id but is location_id
in the second one.

This commit fix the variable name that refers to the destination location

opw : 1919655

closes odoo/odoo#29755