[FIX] web: add tests for calendar start of the week and week numbering
The week_start has been added as an option since commit aa62186855ed48e.
This stores as a database parameter the day starting the week, [1.. 7] starting
on Monday, while the calendar options expect to get Sunday as 7 % 7 = 0.
As a result the week number would be shifted.

Commit 8a8dfe3bf6bdb495c438d8d844c2a1e29d2dbb23 solved the issue,
this commit adds tests for it.

opw 1916666

closes odoo/odoo#29838